corrigindo dto de event

+++ Services/EventService/Dto/EventDto.cs
@@ -7,7 +7,7 @@
         public string Event_title { get; set; }
         public string Event_photo { get; set; }
         public DateTime Event_date { get; set; }
-        public int Event_price { get; set; }
+        public float Event_price { get; set; }
         public string Link_to_buy { get; set; }
         public string Location { get; set; }
         public string Description { get; set; }
